# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "localchat"
  version: "1.0.0"
  title: "localchat"
basePath: "/api/v1"
paths:
  /users:
    get:
      summary: "ユーザ一覧"
      operationId: "usersGET"
      parameters:
      - name: "limit"
        in: "query"
        description: "ページの要素数"
        required: false
        type: "integer"
        format: "int32"
      - name: "offset"
        in: "query"
        description: "ページ開始位置"
        required: false
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ一覧返信"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
    post:
      summary: "ユーザ"
      operationId: "usersPOST"
      parameters:
      - in: "body"
        name: "User"
        description: "User"
        required: false
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
  /users/{user_id}:
    get:
      summary: "１ユーザ取得"
      operationId: "usersUser_idGET"
      parameters:
      - name: "user_id"
        in: "path"
        description: "ユーザID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
    put:
      summary: "１ユーザ更新"
      operationId: "usersUser_idPUT"
      parameters:
      - name: "user_id"
        in: "path"
        description: "ユーザID"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "User"
        description: "User"
        required: false
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
    delete:
      summary: "１ユーザ削除"
      operationId: "usersUser_idDELETE"
      parameters:
      - name: "user_id"
        in: "path"
        description: "ユーザID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
  /users/{user_id}/messages:
    get:
      summary: "ユーザのメッセージ一覧取得"
      operationId: "usersUser_idMessagesGET"
      parameters:
      - name: "user_id"
        in: "path"
        description: "ユーザID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザメッセージ一覧取得"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
  /users/{user_id}/groups:
    get:
      summary: "？？"
      operationId: "usersUser_idGroupsGET"
      parameters:
      - name: "user_id"
        in: "path"
        description: "ユーザID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
    post:
      summary: "ユーザのグループ参加"
      operationId: "usersUser_idGroupsPOST"
      parameters:
      - name: "user_id"
        in: "path"
        description: "ユーザID"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "group"
        description: "group"
        required: false
        schema:
          $ref: "#/definitions/Group"
      responses:
        "200":
          description: "グループ参加"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
  /messages:
    get:
      summary: "メッセージ一覧取得"
      operationId: "messagesGET"
      parameters: []
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
      x-swagger-router-controller: "Default"
    post:
      summary: "メッセージの作成"
      operationId: "messagesPOST"
      parameters:
      - in: "body"
        name: "Message"
        description: "Message"
        required: false
        schema:
          $ref: "#/definitions/Message"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Default"
  /messages/{message_id}:
    get:
      summary: "メッセージの取得"
      operationId: "messagesMessage_idGET"
      parameters:
      - name: "message_id"
        in: "path"
        description: "メッセージID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Default"
    put:
      summary: "メッセージの更新"
      operationId: "messagesMessage_idPUT"
      parameters:
      - name: "message_id"
        in: "path"
        description: "メッセージID"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "message"
        description: "message"
        required: false
        schema:
          $ref: "#/definitions/Message"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Default"
    delete:
      summary: "メッセージの削除"
      operationId: "messagesMessage_idDELETE"
      parameters:
      - name: "message_id"
        in: "path"
        description: "メッセージID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Default"
  /groups:
    get:
      summary: "グループ一覧取得"
      operationId: "groupsGET"
      parameters: []
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
    post:
      summary: "グループ作成"
      operationId: "groupsPOST"
      parameters:
      - in: "body"
        name: "group"
        description: "group"
        required: false
        schema:
          $ref: "#/definitions/Group"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
  /groups/{group_id}:
    get:
      summary: "グループ取得"
      operationId: "groupsGroup_idGET"
      parameters:
      - name: "group_id"
        in: "path"
        description: "グループID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/Group"
      x-swagger-router-controller: "Default"
    put:
      summary: "グループ更新"
      operationId: "groupsGroup_idPUT"
      parameters:
      - name: "group_id"
        in: "path"
        description: "グループID"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "group"
        description: "group"
        required: false
        schema:
          $ref: "#/definitions/Group"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/Group"
      x-swagger-router-controller: "Default"
    delete:
      summary: "グループ削除"
      operationId: "groupsGroup_idDELETE"
      parameters:
      - name: "group_id"
        in: "path"
        description: "グループID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/Group"
      x-swagger-router-controller: "Default"
  /group/{group_id}/members:
    get:
      summary: "グループ内メンバー一覧取得"
      operationId: "groupGroup_idMembersGET"
      parameters:
      - name: "group_id"
        in: "path"
        description: "グループID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Group"
      x-swagger-router-controller: "Default"
  /group/{group_id}/members/{user_id}:
    delete:
      summary: "グループ内のメンバー削除"
      operationId: "groupGroup_idMembersUser_idDELETE"
      parameters:
      - name: "group_id"
        in: "path"
        description: "グループID"
        required: true
        type: "integer"
        format: "int32"
      - name: "user_id"
        in: "path"
        description: "ユーザID"
        required: true
        type: "integer"
        format: "int32"
      responses:
        "200":
          description: "ユーザ返信"
          schema:
            $ref: "#/definitions/User"
      x-swagger-router-controller: "Default"
  /group/{group_id}/messages:
    post:
      summary: "ユーザのメッセージ一送信"
      operationId: "groupGroup_idMessagesPOST"
      parameters:
      - in: "body"
        name: "Message"
        description: "Message"
        required: false
        schema:
          $ref: "#/definitions/Message"
      responses:
        "200":
          description: "ユーザにメッセージ送信"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Default"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      ip_address:
        type: "string"
      pc_name:
        type: "string"
    example:
      pc_name: "pc_name"
      name: "name"
      id: 0
      ip_address: "ip_address"
  Message:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      author:
        type: "integer"
        format: "int32"
      channel:
        type: "integer"
        format: "int32"
      text:
        type: "integer"
        format: "int32"
    example:
      author: 6
      channel: 1
      id: 0
      text: 5
  Group:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
    example:
      name: "name"
      id: 0
  AccessGroup:
    type: "object"
    properties:
      gid:
        type: "integer"
        format: "int32"
      uid:
        type: "integer"
        format: "int32"
  Channel:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      group:
        type: "integer"
        format: "int32"
  GeneralError:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
parameters:
  Limit:
    name: "limit"
    in: "query"
    description: "ページの要素数"
    required: false
    type: "integer"
    format: "int32"
  Offset:
    name: "offset"
    in: "query"
    description: "ページ開始位置"
    required: false
    type: "integer"
    format: "int32"
  From:
    name: "from"
    in: "query"
    description: "期間絞り込み開始日時"
    required: false
    type: "string"
    format: "date-time"
  To:
    name: "to"
    in: "query"
    description: "期間絞り込み終了日時"
    required: false
    type: "string"
    format: "date-time"
  UserId:
    name: "user_id"
    in: "path"
    description: "ユーザID"
    required: true
    type: "integer"
    format: "int32"
  MessageId:
    name: "message_id"
    in: "path"
    description: "メッセージID"
    required: true
    type: "integer"
    format: "int32"
  GroupId:
    name: "group_id"
    in: "path"
    description: "グループID"
    required: true
    type: "integer"
    format: "int32"
  ChannelId:
    name: "channel_id"
    in: "path"
    description: "チャンネルID"
    required: true
    type: "integer"
    format: "int32"